extract sheetOpt interface and constants, valuesFrmattingObj function, minor note updates and edits
